requirements.txt
hidapi
evdev
